\name{ScotCancer}
\docType{data}
\alias{ScotCancer}
\title{Scottish lip cancer dataset from Clayton and Kaldor (1987)}

\format{The data object for the Monte Carlo likelihood estiation and it follows the same structure as described in \code{CAR.simGLM}. The covX is the model matrix with intercept and the fixed effect Paff
}

\description{The Scottish lip cancer dataset.}

\source{Clayton D, Kaldor J 1987. Empirical Bayes Estimation of Age-standardized Relative Risk for use in Disease Mapping. \emph{Biometrics} 43, 671--681
	
Ronnegard L, Shen X, and Alam M 2010 hglm: A Package for Fitting Hierarchical Generalized Linear Models. The R Journal, 2(2): 20-28. URL http://journal.r-project.org/archive/2010-2/RJournal_2010-2_Roennegaard~et~al.pdf.
}

\references{Clayton D, Kaldor J 1987. Empirical Bayes Estimation of Age-standardized Relative Risk for use in Disease Mapping. \emph{Biometrics} 43, 671--681}

\keyword{datasets}