\name{scotplot}
\docType{data}
\alias{scotplot}
\title{Scottish lip cancer dataset from Clayton and Kaldor (1987) with geo information.}
\description{
The Scottish lip cancer dataset with geo information for plotting.
}
\format{
A SpatialPolygonDataFrame contains the geo information and data for plotting
}
\source{Clayton D, Kaldor J 1987. Empirical Bayes Estimation of Age-standardized Relative Risk for use in Disease Mapping. \emph{Biometrics} 43, 671--681

Lee D 2016. CARBayes: Spatial Generalised Linear Mixed Models for Areal Unit Data. R package version 4.6. https://CRAN.R-project.org/package=CARBayes
}
\references{Clayton D, Kaldor J 1987. Empirical Bayes Estimation of Age-standardized Relative Risk for use in Disease Mapping. \emph{Biometrics} 43, 671--681}
\concept{datasets}